[QPG] Introduce BaseApp as a common platform for all QPG apps
Address code review comments
