+++ examples/light-switch-app/qpg/BUILD.gn
@@ -41,6 +41,8 @@ qpg_sdk("sdk") {
   include_dirs = [
     "${chip_root}/src/platform/qpg",
     "${examples_plat_dir}",
+    "${examples_plat_dir}/app",
+    "${examples_plat_dir}/diagnostic_logs",
     "${qpg_project_dir}/include",
   ]
 
@@ -51,9 +53,13 @@ qpg_executable("light_switch_app") {
   output_name = "chip-${qpg_target_ic}-light-switch-example.out"
 
   sources = [
+    "${chip_root}/src/app/clusters/general-diagnostics-server/GenericFaultTestEventTriggerHandler.cpp",
+    "${examples_plat_dir}/app/BaseAppTask.cpp",
     "${examples_plat_dir}/app/battery.cpp",
     "${examples_plat_dir}/app/main.cpp",
+    "${examples_plat_dir}/diagnostic_logs/DiagnosticLogsProviderDelegateImpl.cpp",
     "${examples_plat_dir}/ota/ota.cpp",
+    "${examples_plat_dir}/powercycle_counting.c",
     "${qpg_sdk_root}/Applications/Matter/shared/src/application_header.c",
     "${qpg_sdk_root}/Components/Qorvo/BSP/qButton/src/ButtonHandler.c",
     "${qpg_sdk_root}/Components/Qorvo/BSP/qLed/src/StatusLed.c",
@@ -146,11 +152,6 @@ qpg_executable("light_switch_app") {
 
   ldscript = "${qpg_sdk_root}/Libraries/Qorvo/QorvoStack/gen/QorvoStack_${qpg_target_ic}/QorvoStack_${qpg_target_ic}.ld"
   ldscript = "${qpg_sdk_root}/Applications/Matter/base/gen/base_${qpg_target_ic}_development/base_${qpg_target_ic}_development.ld"
-
-  sources += [
-    "${examples_plat_dir}/diagnostic_logs/DiagnosticLogsProviderDelegateImpl.cpp",
-    "${examples_plat_dir}/diagnostic_logs/DiagnosticLogsProviderDelegateImpl.h",
-  ]
 
   inputs = [ ldscript ]
