+++ examples/light-switch-app/qpg/src/AppTask.cpp
@@ -16,218 +16,77 @@
  *    limitations under the License.
  */
 
-// TODO - Intentionally grouped together
-#include "binding-handler.h"
 #include <lib/support/CHIPMem.h>
-using namespace ::chip;
 
+#include "binding-handler.h"
 #include "gpSched.h"
 
 #include "AppConfig.h"
 #include "AppEvent.h"
 #include "AppTask.h"
+#include "App_Battery.h"
 #include "ota.h"
-
-#include <setup_payload/OnboardingCodesUtil.h>
-
-#include <app/TestEventTriggerDelegate.h>
-#include <app/clusters/general-diagnostics-server/general-diagnostics-server.h>
-#include <app/clusters/identify-server/identify-server.h>
-#include <app/server/Dnssd.h>
-#include <app/server/Server.h>
-#include <app/util/attribute-storage.h>
-#include <data-model-providers/codegen/Instance.h>
-
-#include <credentials/DeviceAttestationCredsProvider.h>
-#include <credentials/examples/DeviceAttestationCredsExample.h>
 
 #include "ButtonHandler.h"
 #include "StatusLed.h"
 #include "qPinCfg.h"
 
-#include <inet/EndPointStateOpenThread.h>
+#if defined(GP_UPGRADE_DIVERSITY_DUAL_BOOT)
+#include "combo-handler.h"
+#endif // GP_UPGRADE_DIVERSITY_DUAL_BOOT
 
-#include <DeviceInfoProviderImpl.h>
-#include <setup_payload/QRCodeSetupPayloadGenerator.h>
-#include <setup_payload/SetupPayload.h>
+#include "DiagnosticLogsProviderDelegateImpl.h"
+#include <app/server/Server.h>
 
 using namespace ::chip;
 using namespace ::chip::app;
 using namespace ::chip::TLV;
 using namespace ::chip::Credentials;
 using namespace ::chip::DeviceLayer;
+using namespace chip::app::Clusters::DiagnosticLogs;
 
 #include <platform/CHIPDeviceLayer.h>
 
-#define FACTORY_RESET_TRIGGER_TIMEOUT 3000
-#define FACTORY_RESET_CANCEL_WINDOW_TIMEOUT 3000
 #define SWITCH_MULTIPRESS_WINDOW_MS 500
 #define SWITCH_LONGPRESS_WINDOW_MS 3000
-#define SWITCH_BUTTON_PRESSED 1
-#define SWITCH_BUTTON_UNPRESSED 0
 
-#define APP_TASK_NAME "APP"
-#define APP_TASK_STACK_SIZE (2 * 1024)
-#define APP_TASK_PRIORITY 2
-#define APP_EVENT_QUEUE_SIZE 10
-#define SECONDS_IN_HOUR (3600)                                              // we better keep this 3600
-#define TOTAL_OPERATIONAL_HOURS_SAVE_INTERVAL_SECONDS (1 * SECONDS_IN_HOUR) // increment every hour
-
-namespace {
-TaskHandle_t sAppTaskHandle;
-QueueHandle_t sAppEventQueue;
-
-bool sIsThreadProvisioned     = false;
-bool sIsThreadAttached        = false;
-bool sHaveBLEConnections      = false;
-bool sIsBLEAdvertisingEnabled = false;
-bool sIsMultipressOngoing     = false;
-bool sLongPressDetected       = false;
-uint8_t sSwitchButtonState    = SWITCH_BUTTON_UNPRESSED;
-
-// NOTE! This key is for test/certification only and should not be available in production devices!
-uint8_t sTestEventTriggerEnableKey[TestEventTriggerDelegate::kEnableKeyLength] = { 0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77,
-                                                                                   0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff };
-
-uint8_t sAppEventQueueBuffer[APP_EVENT_QUEUE_SIZE * sizeof(AppEvent)];
-
-StaticQueue_t sAppEventQueueStruct;
-
-StackType_t appStack[APP_TASK_STACK_SIZE / sizeof(StackType_t)];
-StaticTask_t appTaskStruct;
-
-chip::DeviceLayer::DeviceInfoProviderImpl gExampleDeviceInfoProvider;
+bool sSwitchButtonPressed = false;
+bool sIsMultipressOngoing = false;
+bool sLongPressDetected   = false;
 
 const uint8_t StatusLedGpios[] = QPINCFG_STATUS_LED;
 const ButtonConfig_t buttons[] = QPINCFG_BUTTONS;
-} // namespace
 
 AppTask AppTask::sAppTask;
 
-namespace {
-constexpr int extDiscTimeoutSecs = 20;
-}
-
-Clusters::Identify::EffectIdentifierEnum sIdentifyEffect = Clusters::Identify::EffectIdentifierEnum::kStopEffect;
-
-/**********************************************************
- * Identify Callbacks
- *********************************************************/
-
-namespace {
-void OnTriggerIdentifyEffectCompleted(chip::System::Layer * systemLayer, void * appState)
+CHIP_ERROR AppTask::Init()
 {
-    sIdentifyEffect = Clusters::Identify::EffectIdentifierEnum::kStopEffect;
-}
-} // namespace
-
-void OnTriggerIdentifyEffect(Identify * identify)
-{
-    sIdentifyEffect = identify->mCurrentEffectIdentifier;
-
-    if (identify->mEffectVariant != Clusters::Identify::EffectVariantEnum::kDefault)
+    // Initialize common code in base class
+    CHIP_ERROR err = BaseAppTask::Init();
+    if (err != CHIP_NO_ERROR)
     {
-        ChipLogDetail(AppServer, "Identify Effect Variant unsupported. Using default");
+        ChipLogError(NotSpecified, "BaseAppTask::Init() failed");
+        return err;
     }
 
-    switch (sIdentifyEffect)
-    {
-    case Clusters::Identify::EffectIdentifierEnum::kBlink:
-    case Clusters::Identify::EffectIdentifierEnum::kBreathe:
-    case Clusters::Identify::EffectIdentifierEnum::kOkay:
-    case Clusters::Identify::EffectIdentifierEnum::kChannelChange:
-        SystemLayer().ScheduleLambda([identify] {
-            (void) chip::DeviceLayer::SystemLayer().StartTimer(chip::System::Clock::Seconds16(5), OnTriggerIdentifyEffectCompleted,
-                                                               identify);
-        });
-        break;
-    case Clusters::Identify::EffectIdentifierEnum::kFinishEffect:
-        SystemLayer().ScheduleLambda([identify] {
-            (void) chip::DeviceLayer::SystemLayer().CancelTimer(OnTriggerIdentifyEffectCompleted, identify);
-            (void) chip::DeviceLayer::SystemLayer().StartTimer(chip::System::Clock::Seconds16(1), OnTriggerIdentifyEffectCompleted,
-                                                               identify);
-        });
-        break;
-    case Clusters::Identify::EffectIdentifierEnum::kStopEffect:
-        SystemLayer().ScheduleLambda(
-            [identify] { (void) chip::DeviceLayer::SystemLayer().CancelTimer(OnTriggerIdentifyEffectCompleted, identify); });
-        sIdentifyEffect = Clusters::Identify::EffectIdentifierEnum::kStopEffect;
-        break;
-    default:
-        ChipLogProgress(Zcl, "No identifier effect");
-    }
-}
+    // Init ZCL Data Model and start server
+    PlatformMgr().ScheduleWork(AppTask::InitServerWrapper, 0);
 
-Identify gIdentifyEp1 = {
-    chip::EndpointId{ 1 },
-    [](Identify *) { ChipLogProgress(Zcl, "onIdentifyStart"); },
-    [](Identify *) { ChipLogProgress(Zcl, "onIdentifyStop"); },
-    Clusters::Identify::IdentifyTypeEnum::kNone,
-    OnTriggerIdentifyEffect,
-};
+    // Setup button handler
+    ButtonHandler_Init(buttons, Q_ARRAY_SIZE(buttons), BUTTON_LOW, AppTask::ButtonEventHandlerWrapper);
 
-Identify gIdentifyEp2 = {
-    chip::EndpointId{ 2 },
-    [](Identify *) { ChipLogProgress(Zcl, "onIdentifyStart"); },
-    [](Identify *) { ChipLogProgress(Zcl, "onIdentifyStop"); },
-    Clusters::Identify::IdentifyTypeEnum::kNone,
-    OnTriggerIdentifyEffect,
-};
+    PlatformMgr().ScheduleWork(InitBindingManager);
 
-void LockOpenThreadTask(void)
-{
-    chip::DeviceLayer::ThreadStackMgr().LockThreadStack();
-}
+    SwitchMgr().Init();
 
-void UnlockOpenThreadTask(void)
-{
-    chip::DeviceLayer::ThreadStackMgr().UnlockThreadStack();
-}
+    gpSched_ScheduleEvent(1 * ONE_SECOND_US, Application_StartPeriodicBatteryUpdate);
 
-CHIP_ERROR AppTask::StartAppTask()
-{
-    sAppEventQueue = xQueueCreateStatic(APP_EVENT_QUEUE_SIZE, sizeof(AppEvent), sAppEventQueueBuffer, &sAppEventQueueStruct);
-    if (sAppEventQueue == nullptr)
-    {
-        ChipLogError(NotSpecified, "Failed to allocate app event queue");
-        return CHIP_ERROR_NO_MEMORY;
-    }
-
-    // Start App task.
-    sAppTaskHandle =
-        xTaskCreateStatic(AppTaskMain, APP_TASK_NAME, MATTER_ARRAY_SIZE(appStack), nullptr, 1, appStack, &appTaskStruct);
-    if (sAppTaskHandle == nullptr)
-    {
-        return CHIP_ERROR_NO_MEMORY;
-    }
-
-    return CHIP_NO_ERROR;
+    return err;
 }
 
 void AppTask::InitServer(intptr_t arg)
 {
-    static chip::CommonCaseDeviceServerInitParams initParams;
-
-    chip::Inet::EndPointStateOpenThread::OpenThreadEndpointInitParam nativeParams;
-    nativeParams.lockCb                = LockOpenThreadTask;
-    nativeParams.unlockCb              = UnlockOpenThreadTask;
-    nativeParams.openThreadInstancePtr = chip::DeviceLayer::ThreadStackMgrImpl().OTInstance();
-    initParams.endpointNativeParams    = static_cast<void *>(&nativeParams);
-
-    static SimpleTestEventTriggerDelegate sTestEventTriggerDelegate{};
-    VerifyOrDie(sTestEventTriggerDelegate.Init(ByteSpan(sTestEventTriggerEnableKey)) == CHIP_NO_ERROR);
-    (void) initParams.InitializeStaticResourcesBeforeServerInit();
-    initParams.dataModelProvider        = CodegenDataModelProviderInstance(initParams.persistentStorageDelegate);
-    initParams.testEventTriggerDelegate = &sTestEventTriggerDelegate;
-
-    gExampleDeviceInfoProvider.SetStorageDelegate(initParams.persistentStorageDelegate);
-    chip::DeviceLayer::SetDeviceInfoProvider(&gExampleDeviceInfoProvider);
-
-    chip::Server::GetInstance().Init(initParams);
-
-#if CHIP_DEVICE_CONFIG_ENABLE_EXTENDED_DISCOVERY
-    chip::app::DnssdServer::Instance().SetExtendedDiscoveryTimeoutSecs(extDiscTimeoutSecs);
-#endif
+    BaseAppTask::InitServer(arg);
 
     // Open commissioning after boot if no fabric was available
     if (chip::Server::GetInstance().GetFabricTable().FabricCount() == 0)
@@ -236,83 +95,21 @@ void AppTask::InitServer(intptr_t arg)
     }
 }
 
-void AppTask::OpenCommissioning(intptr_t arg)
+bool AppTask::ButtonEventHandler(uint8_t btnIdx, bool btnPressed)
 {
-    // Enable BLE advertisements
-    chip::Server::GetInstance().GetCommissioningWindowManager().OpenBasicCommissioningWindow();
-    ChipLogProgress(NotSpecified, "BLE advertising started. Waiting for Pairing.");
-}
-
-CHIP_ERROR AppTask::Init()
-{
-    CHIP_ERROR err      = CHIP_NO_ERROR;
-    const qResult_t res = qPinCfg_Init(NULL);
-
-    if (res != Q_OK)
+    // Call base class ButtonEventHandler
+    bool eventHandled = BaseAppTask::ButtonEventHandler(btnIdx, btnPressed);
+    if (eventHandled)
     {
-        ChipLogError(NotSpecified, "qPinCfg_Init failed: %d", res);
+        return true;
     }
 
-    StatusLed_Init(StatusLedGpios, Q_ARRAY_SIZE(StatusLedGpios), true);
-
-    PlatformMgr().AddEventHandler(MatterEventHandler, 0);
-
-    ChipLogProgress(NotSpecified, "Current Software Version: %s", CHIP_DEVICE_CONFIG_DEVICE_SOFTWARE_VERSION_STRING);
-
-    // Init ZCL Data Model and start server
-    PlatformMgr().ScheduleWork(InitServer, 0);
-
-    ReturnErrorOnFailure(mFactoryDataProvider.Init());
-    SetDeviceInstanceInfoProvider(&mFactoryDataProvider);
-    SetCommissionableDataProvider(&mFactoryDataProvider);
-
-    SetDeviceAttestationCredentialsProvider(&mFactoryDataProvider);
-
-    UpdateClusterState();
-
-    // Setup button handler
-    ButtonHandler_Init(buttons, Q_ARRAY_SIZE(buttons), BUTTON_LOW, ButtonEventHandler);
-
-    DeviceLayer::PlatformMgr().ScheduleWork(InitBindingManager);
-
-    // Log device configuration
-    ConfigurationMgr().LogDeviceConfig();
-    PrintOnboardingCodes(chip::RendezvousInformationFlags(chip::RendezvousInformationFlag::kBLE));
-
-    sIsThreadProvisioned     = ConnectivityMgr().IsThreadProvisioned();
-    sIsThreadAttached        = ConnectivityMgr().IsThreadAttached();
-    sHaveBLEConnections      = (ConnectivityMgr().NumBLEConnections() != 0);
-    sIsBLEAdvertisingEnabled = ConnectivityMgr().IsBLEAdvertisingEnabled();
-    UpdateLEDs();
-
-    err = chip::DeviceLayer::SystemLayer().StartTimer(chip::System::Clock::Seconds32(TOTAL_OPERATIONAL_HOURS_SAVE_INTERVAL_SECONDS),
-                                                      TotalHoursTimerHandler, this);
-
-    if (err != CHIP_NO_ERROR)
+    // Only go ahead if button index has a supported value
+    if (btnIdx != APP_LEVEL_BUTTON && btnIdx != APP_SWITCH_BUTTON && btnIdx != APP_COLOR_BUTTON && btnIdx != APP_TOGGLE_BUTTON)
     {
-        ChipLogError(NotSpecified, "StartTimer failed %s: ", chip::ErrorStr(err));
+        return false;
     }
 
-    return err;
-}
-
-void AppTask::AppTaskMain(void * pvParameter)
-{
-    AppEvent event;
-
-    while (true)
-    {
-        BaseType_t eventReceived = xQueueReceive(sAppEventQueue, &event, portMAX_DELAY);
-        while (eventReceived == pdTRUE)
-        {
-            sAppTask.DispatchEvent(&event);
-            eventReceived = xQueueReceive(sAppEventQueue, &event, 0);
-        }
-    }
-}
-
-void AppTask::ButtonEventHandler(uint8_t btnIdx, bool btnPressed)
-{
     CHIP_ERROR err = CHIP_NO_ERROR;
 
     ChipLogDetail(NotSpecified, "ButtonEventHandler %d, %d", btnIdx, btnPressed);
@@ -322,19 +119,13 @@ void AppTask::ButtonEventHandler(uint8_t btnIdx, bool btnPressed)
     button_event.ButtonEvent.ButtonIdx = btnIdx;
     button_event.ButtonEvent.Action    = btnPressed;
 
-    switch (btnIdx)
+    if (btnIdx == APP_LEVEL_BUTTON && btnPressed == true)
     {
-    case APP_LEVEL_BUTTON: {
-        if (!btnPressed)
-        {
-            return;
-        }
-
         ChipLogProgress(NotSpecified, "Level Button pressed");
         button_event.Handler = SwitchMgr().LevelHandler;
-        break;
     }
-    case APP_SWITCH_BUTTON: {
+    else if (btnIdx == APP_SWITCH_BUTTON)
+    {
         if (!btnPressed)
         {
             ChipLogDetail(NotSpecified, "Switch button released");
@@ -343,7 +134,7 @@ void AppTask::ButtonEventHandler(uint8_t btnIdx, bool btnPressed)
                 sLongPressDetected ? SwitchMgr().GenericSwitchLongReleaseHandler : SwitchMgr().GenericSwitchShortReleaseHandler;
 
             sIsMultipressOngoing = true;
-            sSwitchButtonState   = SWITCH_BUTTON_UNPRESSED;
+            sSwitchButtonPressed = false;
             sLongPressDetected   = false;
 
             chip::DeviceLayer::SystemLayer().CancelTimer(MultiPressTimeoutHandler, NULL);
@@ -360,7 +151,7 @@ void AppTask::ButtonEventHandler(uint8_t btnIdx, bool btnPressed)
         {
             ChipLogDetail(NotSpecified, "Switch button pressed");
 
-            sSwitchButtonState = SWITCH_BUTTON_PRESSED;
+            sSwitchButtonPressed = true;
 
             chip::DeviceLayer::SystemLayer().CancelTimer(LongPressTimeoutHandler, NULL);
             // we need to check if this is short or long press
@@ -386,47 +177,25 @@ void AppTask::ButtonEventHandler(uint8_t btnIdx, bool btnPressed)
                 button_event.Handler = SwitchMgr().GenericSwitchInitialPressHandler;
             }
         }
-        break;
     }
-    case APP_COLOR_BUTTON: {
-        if (!btnPressed)
-        {
-            return;
-        }
+    else if (btnIdx == APP_COLOR_BUTTON && btnPressed == true)
+    {
         ChipLogProgress(NotSpecified, "Color button pressed");
         button_event.Handler = SwitchMgr().ColorHandler;
-        break;
     }
-    case APP_TOGGLE_BUTTON: {
-        if (!btnPressed)
-        {
-            return;
-        }
-
+    else if (btnIdx == APP_TOGGLE_BUTTON && btnPressed == true)
+    {
         ChipLogProgress(NotSpecified, "Toggle Button pressed");
         button_event.Handler = SwitchMgr().ToggleHandler;
-        break;
     }
-    case APP_FUNCTION_BUTTON: {
-        button_event.Handler = FunctionHandler;
-        break;
-    }
-    default: {
-        // invalid button
-        return;
-    }
+    else
+    {
+        return false;
     }
 
     sAppTask.PostEvent(&button_event);
-}
 
-void AppTask::TimerEventHandler(chip::System::Layer * aLayer, void * aAppState)
-{
-    AppEvent event;
-    event.Type               = AppEvent::kEventType_Timer;
-    event.TimerEvent.Context = aAppState;
-    event.Handler            = FunctionTimerEventHandler;
-    sAppTask.PostEvent(&event);
+    return true;
 }
 
 void AppTask::MultiPressTimeoutHandler(chip::System::Layer * aLayer, void * aAppState)
@@ -447,7 +216,7 @@ void AppTask::LongPressTimeoutHandler(chip::System::Layer * aLayer, void * aAppS
     ChipLogDetail(NotSpecified, "LongPressTimeoutHandler");
 
     // if the button is still pressed after threshold time, this is a LongPress, otherwise jsut ignore it
-    if (sSwitchButtonState == SWITCH_BUTTON_PRESSED && sIsMultipressOngoing == false)
+    if (sSwitchButtonPressed == true && sIsMultipressOngoing == false)
     {
         sLongPressDetected       = true;
         AppEvent longpress_event = {};
@@ -455,270 +224,5 @@ void AppTask::LongPressTimeoutHandler(chip::System::Layer * aLayer, void * aAppS
         longpress_event.Handler  = SwitchMgr().GenericSwitchLongPressHandler;
 
         sAppTask.PostEvent(&longpress_event);
-    }
-}
-
-void AppTask::TotalHoursTimerHandler(chip::System::Layer * aLayer, void * aAppState)
-{
-    ChipLogDetail(NotSpecified, "HourlyTimer");
-
-    CHIP_ERROR err;
-    uint32_t totalOperationalHours = 0;
-
-    err = ConfigurationMgr().GetTotalOperationalHours(totalOperationalHours);
-
-    if (err == CHIP_NO_ERROR)
-    {
-        ConfigurationMgr().StoreTotalOperationalHours(totalOperationalHours +
-                                                      (TOTAL_OPERATIONAL_HOURS_SAVE_INTERVAL_SECONDS / SECONDS_IN_HOUR));
-    }
-    else if (err == CHIP_DEVICE_ERROR_CONFIG_NOT_FOUND)
-    {
-        totalOperationalHours = 0; // set this explicitly to 0 for safety
-        ConfigurationMgr().StoreTotalOperationalHours(totalOperationalHours +
-                                                      (TOTAL_OPERATIONAL_HOURS_SAVE_INTERVAL_SECONDS / SECONDS_IN_HOUR));
-    }
-    else
-    {
-        ChipLogError(DeviceLayer, "Failed to get total operational hours of the Node");
-    }
-
-    err = chip::DeviceLayer::SystemLayer().StartTimer(chip::System::Clock::Seconds32(TOTAL_OPERATIONAL_HOURS_SAVE_INTERVAL_SECONDS),
-                                                      TotalHoursTimerHandler, nullptr);
-
-    if (err != CHIP_NO_ERROR)
-    {
-        ChipLogError(NotSpecified, "StartTimer failed %s: ", chip::ErrorStr(err));
-    }
-}
-
-void AppTask::FunctionTimerEventHandler(AppEvent * aEvent)
-{
-    if (aEvent->Type != AppEvent::kEventType_Timer)
-    {
-        return;
-    }
-
-    // If we reached here, the button was held past FACTORY_RESET_TRIGGER_TIMEOUT,
-    // initiate factory reset
-    if (sAppTask.mFunctionTimerActive && sAppTask.mFunction == kFunction_SoftwareUpdate)
-    {
-        ChipLogProgress(NotSpecified, "[BTN] Factory Reset selected. Release within %us to cancel.",
-                        FACTORY_RESET_CANCEL_WINDOW_TIMEOUT / 1000);
-
-        // Start timer for FACTORY_RESET_CANCEL_WINDOW_TIMEOUT to allow user to
-        // cancel, if required.
-        sAppTask.StartTimer(FACTORY_RESET_CANCEL_WINDOW_TIMEOUT);
-
-        sAppTask.mFunction = kFunction_FactoryReset;
-
-        // Turn off all LEDs before starting blink to make sure blink is
-        // co-ordinated.
-        StatusLed_SetLed(SYSTEM_STATE_LED, false);
-
-        StatusLed_BlinkLed(SYSTEM_STATE_LED, 500, 500);
-    }
-    else if (sAppTask.mFunctionTimerActive && sAppTask.mFunction == kFunction_FactoryReset)
-    {
-        // Actually trigger Factory Reset
-        sAppTask.mFunction = kFunction_NoneSelected;
-        chip::Server::GetInstance().ScheduleFactoryReset();
-    }
-}
-
-void AppTask::FunctionHandler(AppEvent * aEvent)
-{
-    if (aEvent->ButtonEvent.ButtonIdx != APP_FUNCTION_BUTTON)
-    {
-        return;
-    }
-
-    // To trigger software update: press the APP_FUNCTION_BUTTON button briefly (<
-    // FACTORY_RESET_TRIGGER_TIMEOUT) To initiate factory reset: press the
-    // APP_FUNCTION_BUTTON for FACTORY_RESET_TRIGGER_TIMEOUT +
-    // FACTORY_RESET_CANCEL_WINDOW_TIMEOUT All LEDs start blinking after
-    // FACTORY_RESET_TRIGGER_TIMEOUT to signal factory reset has been initiated.
-    // To cancel factory reset: release the APP_FUNCTION_BUTTON once all LEDs
-    // start blinking within the FACTORY_RESET_CANCEL_WINDOW_TIMEOUT
-    if (aEvent->ButtonEvent.Action == true)
-    {
-        if (!sAppTask.mFunctionTimerActive && sAppTask.mFunction == kFunction_NoneSelected)
-        {
-            ChipLogProgress(NotSpecified, "[BTN] Hold to select function:");
-            ChipLogProgress(NotSpecified, "[BTN] - Trigger OTA (0-3s)");
-            ChipLogProgress(NotSpecified, "[BTN] - Factory Reset (>6s)");
-
-            sAppTask.StartTimer(FACTORY_RESET_TRIGGER_TIMEOUT);
-
-            sAppTask.mFunction = kFunction_SoftwareUpdate;
-        }
-    }
-    else
-    {
-        // If the button was released before factory reset got initiated, trigger a
-        // software update.
-        if (sAppTask.mFunctionTimerActive && sAppTask.mFunction == kFunction_SoftwareUpdate)
-        {
-            sAppTask.CancelTimer();
-
-            sAppTask.mFunction = kFunction_NoneSelected;
-
-            ChipLogProgress(NotSpecified, "[BTN] Triggering OTA Query");
-
-            TriggerOTAQuery();
-        }
-        else if (sAppTask.mFunctionTimerActive && sAppTask.mFunction == kFunction_FactoryReset)
-        {
-            sAppTask.CancelTimer();
-
-            // Change the function to none selected since factory reset has been
-            // canceled.
-            sAppTask.mFunction = kFunction_NoneSelected;
-
-            ChipLogProgress(NotSpecified, "[BTN] Factory Reset has been Canceled");
-        }
-    }
-}
-
-void AppTask::CancelTimer()
-{
-    chip::DeviceLayer::SystemLayer().CancelTimer(TimerEventHandler, this);
-    mFunctionTimerActive = false;
-}
-
-void AppTask::StartTimer(uint32_t aTimeoutInMs)
-{
-    CHIP_ERROR err;
-
-    chip::DeviceLayer::SystemLayer().CancelTimer(TimerEventHandler, this);
-    err = chip::DeviceLayer::SystemLayer().StartTimer(chip::System::Clock::Milliseconds32(aTimeoutInMs), TimerEventHandler, this);
-    SuccessOrExit(err);
-
-    mFunctionTimerActive = true;
-exit:
-    if (err != CHIP_NO_ERROR)
-    {
-        ChipLogError(NotSpecified, "StartTimer failed %s: ", chip::ErrorStr(err));
-    }
-}
-
-void AppTask::PostEvent(const AppEvent * aEvent)
-{
-    if (sAppEventQueue != nullptr)
-    {
-        if (!xQueueSend(sAppEventQueue, aEvent, 1))
-        {
-            ChipLogError(NotSpecified, "Failed to post event to app task event queue");
-        }
-    }
-    else
-    {
-        ChipLogError(NotSpecified, "Event Queue is nullptr should never happen");
-    }
-}
-
-void AppTask::DispatchEvent(AppEvent * aEvent)
-{
-    if (aEvent->Handler)
-    {
-        aEvent->Handler(aEvent);
-    }
-    else
-    {
-        ChipLogError(NotSpecified, "Event received with no handler. Dropping event.");
-    }
-}
-
-/**
- * Update cluster status after application level changes
- */
-void AppTask::UpdateClusterState(void)
-{
-    ChipLogProgress(NotSpecified, "UpdateClusterState");
-
-    /* write the new attribute value based on attribute setter API.
-       example API usage of on-off attribute:
-
-        Protocols::InteractionModel::Status status = Clusters::OnOff::Attributes::OnOff::Set(QPG_LIGHT_ENDPOINT_ID,
-       LightingMgr().IsTurnedOn());
-
-        if (status != Protocols::InteractionModel::Status::Success)
-        {
-            ChipLogError(NotSpecified, "ERR: updating on/off %x", to_underlying(status));
-        }
-    */
-}
-
-void AppTask::UpdateLEDs(void)
-{
-    // If system has "full connectivity", keep the LED On constantly.
-    //
-    // If thread attached to the thread network and service provisioned
-    // OR no connectivity to the service OR subscriptions are not fully
-    // established THEN blink the LED Off for a short period of time.
-    //
-    // If the system has ble connection(s) uptill the stage above, THEN blink
-    // the LEDs at an even rate of 100ms.
-    //
-    // Otherwise, turn the LED OFF.
-    if (sIsThreadProvisioned && sIsThreadAttached)
-    {
-        StatusLed_SetLed(SYSTEM_STATE_LED, true);
-    }
-    else if (sIsThreadProvisioned && !sIsThreadAttached)
-    {
-        StatusLed_BlinkLed(SYSTEM_STATE_LED, 950, 50);
-    }
-    else if (sHaveBLEConnections)
-    {
-        StatusLed_BlinkLed(SYSTEM_STATE_LED, 100, 100);
-    }
-    else if (sIsBLEAdvertisingEnabled)
-    {
-        StatusLed_BlinkLed(SYSTEM_STATE_LED, 50, 50);
-    }
-    else
-    {
-        // not commissioned yet
-        StatusLed_SetLed(SYSTEM_STATE_LED, false);
-    }
-}
-
-void AppTask::MatterEventHandler(const ChipDeviceEvent * event, intptr_t)
-{
-    switch (event->Type)
-    {
-    case DeviceEventType::kServiceProvisioningChange: {
-        sIsThreadProvisioned = event->ServiceProvisioningChange.IsServiceProvisioned;
-        UpdateLEDs();
-        break;
-    }
-
-    case DeviceEventType::kThreadConnectivityChange: {
-        sIsThreadAttached = (event->ThreadConnectivityChange.Result == kConnectivity_Established);
-        UpdateLEDs();
-        break;
-    }
-
-    case DeviceEventType::kCHIPoBLEConnectionEstablished: {
-        sHaveBLEConnections = true;
-        UpdateLEDs();
-        break;
-    }
-
-    case DeviceEventType::kCHIPoBLEConnectionClosed: {
-        sHaveBLEConnections = false;
-        UpdateLEDs();
-        break;
-    }
-
-    case DeviceEventType::kCHIPoBLEAdvertisingChange: {
-        sIsBLEAdvertisingEnabled = (event->CHIPoBLEAdvertisingChange.Result == kActivity_Started);
-        UpdateLEDs();
-        break;
-    }
-
-    default:
-        break;
     }
 }
