+++ examples/light-switch-app/qpg/src/ZclCallbacks.cpp
@@ -28,11 +28,24 @@
 #include <assert.h>
 #include <lib/support/logging/CHIPLogging.h>
 
+#include "diagnostic_logs/DiagnosticLogsProviderDelegateImpl.h"
+#include <app/clusters/diagnostic-logs-server/diagnostic-logs-server.h>
+
 using namespace ::chip;
 using namespace chip::app::Clusters;
+
+using namespace ::chip::app::Clusters::DiagnosticLogs;
 
 void MatterPostAttributeChangeCallback(const chip::app::ConcreteAttributePath & attributePath, uint8_t type, uint16_t size,
                                        uint8_t * value)
 {
     return;
+}
+
+void emberAfDiagnosticLogsClusterInitCallback(chip::EndpointId endpoint)
+{
+    ChipLogProgress(NotSpecified, "Setting log provider.");
+
+    auto & logProvider = LogProvider::GetInstance();
+    DiagnosticLogsServer::Instance().SetDiagnosticLogsProviderDelegate(endpoint, &logProvider);
 }
