+++ examples/lighting-app/qpg/include/CHIPProjectConfig.h
@@ -166,3 +166,7 @@
 #endif
 
 #define CHIP_DEVICE_CONFIG_ENABLE_EXTENDED_DISCOVERY 1
+
+#ifndef CHIP_DEVICE_CONFIG_MAX_DIAG_LOG_SIZE
+#define CHIP_DEVICE_CONFIG_MAX_DIAG_LOG_SIZE 2048
+#endif // CHIP_DEVICE_CONFIG_MAX_DIAG_LOG_SIZE
