+++ examples/lock-app/qpg/BUILD.gn
@@ -41,6 +41,8 @@ qpg_sdk("sdk") {
   include_dirs = [
     "${chip_root}/src/platform/qpg",
     "${examples_plat_dir}",
+    "${examples_plat_dir}/app",
+    "${examples_plat_dir}/diagnostic_logs",
     "${qpg_project_dir}/include",
   ]
 
@@ -51,7 +53,10 @@ qpg_executable("lock_app") {
   output_name = "chip-${qpg_target_ic}-lock-example.out"
 
   sources = [
+    "${chip_root}/src/app/clusters/general-diagnostics-server/GenericFaultTestEventTriggerHandler.cpp",
+    "${examples_plat_dir}/app/BaseAppTask.cpp",
     "${examples_plat_dir}/app/main.cpp",
+    "${examples_plat_dir}/diagnostic_logs/DiagnosticLogsProviderDelegateImpl.cpp",
     "${examples_plat_dir}/ota/ota.cpp",
     "${examples_plat_dir}/powercycle_counting.c",
     "${qpg_sdk_root}/Applications/Matter/shared/src/application_header.c",
