+++ examples/lock-app/qpg/include/CHIPProjectConfig.h
@@ -138,8 +138,6 @@
  */
 #define CHIP_IM_MAX_NUM_WRITE_CLIENT 2
 
-#define CHIP_DEVICE_CONFIG_ENABLE_EXTENDED_DISCOVERY 1
-
 /**
  * @def CHIP_DEVICE_CONFIG_ENABLE_SED
  *
@@ -171,3 +169,9 @@
 #ifndef CHIP_CONFIG_ENABLE_ACL_EXTENSIONS
 #define CHIP_CONFIG_ENABLE_ACL_EXTENSIONS 1
 #endif
+
+#define CHIP_DEVICE_CONFIG_ENABLE_EXTENDED_DISCOVERY 1
+
+#ifndef CHIP_DEVICE_CONFIG_MAX_DIAG_LOG_SIZE
+#define CHIP_DEVICE_CONFIG_MAX_DIAG_LOG_SIZE 2048
+#endif // CHIP_DEVICE_CONFIG_MAX_DIAG_LOG_SIZE
