+++ examples/thermostat/qpg/BUILD.gn
@@ -41,6 +41,8 @@ qpg_sdk("sdk") {
   include_dirs = [
     "${chip_root}/src/platform/qpg",
     "${examples_plat_dir}",
+    "${examples_plat_dir}/app",
+    "${examples_plat_dir}/diagnostic_logs",
     "${qpg_project_dir}/include",
   ]
 
@@ -51,8 +53,12 @@ qpg_executable("thermostat") {
   output_name = "chip-${qpg_target_ic}-thermostat-example.out"
 
   sources = [
+    "${chip_root}/examples/thermostat/thermostat-common/src/thermostat-delegate-impl.cpp",
+    "${chip_root}/src/app/clusters/general-diagnostics-server/GenericFaultTestEventTriggerHandler.cpp",
+    "${examples_plat_dir}/app/BaseAppTask.cpp",
     "${examples_plat_dir}/app/battery.cpp",
     "${examples_plat_dir}/app/main.cpp",
+    "${examples_plat_dir}/diagnostic_logs/DiagnosticLogsProviderDelegateImpl.cpp",
     "${examples_plat_dir}/ota/ota.cpp",
     "${examples_plat_dir}/powercycle_counting.c",
     "${qpg_sdk_root}/Applications/Matter/shared/src/application_header.c",
@@ -86,6 +92,7 @@ qpg_executable("thermostat") {
   include_dirs = [
     "include",
     "${examples_plat_dir}/ota",
+    "${chip_root}/examples/thermostat/thermostat-common/include",
     "${qpg_sdk_root}/Components/Qorvo/HAL_PLATFORM/halCortexM4/inc",
     "${qpg_sdk_root}/Components/ThirdParty/secure_element/software/esec/libs/common",
     "${qpg_sdk_root}/Components/ThirdParty/secure_element/software/esec/libs/base/include",  # esec_reatures.h
