+++ examples/thermostat/qpg/include/AppConfig.h
@@ -16,7 +16,7 @@
  *    limitations under the License.
  */
 
-#ifndef APP_CONFIG_H
-#define APP_CONFIG_H
+#pragma once
 
-#endif // APP_CONFIG_H
+#define BASE_APP_BUILD 0
+#define ACTIVATE_IDENTIFY_ON_EP2 0
